Fix handling of task parameter for TL and non-TL models
The previous implementation woudl add an additional task column for
data passed to the TL models, which contraticted with the original
task column from the search space and lead to no source data used
for training. The task parameter is now handled correctly by the
search spaces passed to the models.